add dominates check

+++ Games/types/Mafia/roles/cards/JailTarget.js
@@ -35,6 +35,7 @@ module.exports = class JailTarget extends Card {
                     return false;
                 },
                 action: {
+                    labels: ["kill", "jail"],
                     priority: PRIORITY_JAIL_EXECUTE,
                     run: function () {
                         var prisoner = this.actor.role.data.prisoner;
@@ -42,7 +43,7 @@ module.exports = class JailTarget extends Card {
                         if (!prisoner)
                             return;
 
-                        if (this.target == "Yes")
+                        if (this.target == "Yes" && this.dominates(prisoner)) 
                             prisoner.kill("basic", this.actor);
                     }
                 }
